+++ NEWS.md
@@ -1,3 +1,10 @@
+enmSdmX 1.0.1
+===========
+o Fixed bug experienced by some users using predictEnmSdm() and predictMaxNet() (Thank you, Nikki!)
+o Fixed bug in trainByCrossValid() using improper call to evalContBoyce()
+o Fixed extract() bug in some examples
+o summarizeByCrossValid() now summarizes natural spline (NS) models
+
 enmSdmX 1.0.0
 ===========
 o First release on CRAN